+++ x/vm/keeper/grpc_query.go
@@ -717,8 +717,17 @@ func (k Keeper) TraceCall(c context.Context, req *types.QueryTraceCallRequest) (
 	}
 	msg := args.ToMessage(baseFee, true, true)
 
+	// Get state overrides
+	var stateOverrides *rpctypes.StateOverride
+	if len(req.Overrides) > 0 {
+		stateOverrides = new(rpctypes.StateOverride)
+		if err = json.Unmarshal(req.Overrides, stateOverrides); err != nil {
+			return nil, status.Error(codes.InvalidArgument, fmt.Sprintf("invalid state overrides format: %s", err.Error()))
+		}
+	}
+
 	// trace call
-	result, _, err := k.traceTxWithMsg(ctx, cfg, txConfig, msg, req.GetTraceConfig(), false)
+	result, _, err := k.traceTxWithMsg(ctx, cfg, txConfig, msg, req.GetTraceConfig(), false, stateOverrides)
 	if err != nil {
 		// error will be returned with detail status from traceTx
 		return nil, err
@@ -749,7 +758,7 @@ func (k *Keeper) traceTx(
 		return nil, 0, status.Error(codes.Internal, err.Error())
 	}
 
-	return k.traceTxWithMsg(ctx, cfg, txConfig, msg, traceConfig, commitMessage)
+	return k.traceTxWithMsg(ctx, cfg, txConfig, msg, traceConfig, commitMessage, nil)
 }
 
 // traceTxWithMsg do trace on one Ethereum message, it returns a tuple: (traceResult, nextLogIndex, error).
@@ -760,6 +769,7 @@ func (k *Keeper) traceTxWithMsg(
 	msg *core.Message,
 	traceConfig *types.TraceConfig,
 	commitMessage bool,
+	stateOverrides *rpctypes.StateOverride,
 ) (*interface{}, uint, error) {
 	// Assemble the structured logger or the JavaScript tracer
 	var (
@@ -837,7 +847,7 @@ func (k *Keeper) traceTxWithMsg(
 	// Build EVM execution context
 	ctx = buildTraceCtx(ctx, msg.GasLimit)
 	stateDB := statedb.New(ctx, k, txConfig)
-	res, err := k.ApplyMessageWithConfig(ctx, stateDB, *msg, tracer.Hooks, commitMessage, false, cfg, txConfig, false, nil)
+	res, err := k.ApplyMessageWithConfig(ctx, stateDB, *msg, tracer.Hooks, commitMessage, false, cfg, txConfig, false, stateOverrides)
 	if err != nil {
 		return nil, 0, status.Error(codes.Internal, err.Error())
 	}